{{ cookiecutter.project_slug }}/{{ cookiecutter.project_slug }}/{{ cookiecutter.app_name }}/management/commands/foo_command.py
from django.conf import settings
from django.core.management.base import BaseCommand


import structlog

logger = structlog.get_logger()


class Command(BaseCommand):
    help = 'Basic Command'

    def handle(self, *args, **options):
       logger.info("I'm a management command")
