{{ cookiecutter.project_slug }}/requirements.txt
#### EDITOR ###
pylint==3.0.3
ruff==0.2.2
mypy==1.8.0
django-stubs
pre-commit==3.6.2
pylint-django==2.5.5

### TESTING ###
coverage==7.4.4
pytest==8.1.1
freezegun==1.4.0


### LOCAL ###
ipython==8.22.1
Werkzeug==3.0.1
madbg==1.3.2
ipdb==0.13.13


### PROD ###
Django==5.0.2
djangorestframework==3.14.0
django-extensions==3.2.3
django-cors-headers==4.3.1
django-structlog==7.1.0

psycopg[binary]==3.1.17
gunicorn==21.2.0
pydantic==2.6.2
structlog-sentry==2.1.0
sentry-sdk==1.44.1
colorama==0.4.6
rich==13.7.1
PyYAML==6.0.1
